+++ lib/src/main/java/org/altbeacon/beacon/startup/StartupBroadcastReceiver.java
@@ -35,7 +35,7 @@ public class StartupBroadcastReceiver extends BroadcastReceiver
             int bleCallbackType = intent.getIntExtra(BluetoothLeScanner.EXTRA_CALLBACK_TYPE, -1); // e.g. ScanSettings.CALLBACK_TYPE_FIRST_MATCH
             if (bleCallbackType != -1) {
                 LogManager.d(TAG, "Passive background scan callback type: "+bleCallbackType);
-                LogManager.d(TAG, "got Android O background scan via intent");
+                LogManager.d(TAG, "got Android background scan via intent");
                 int errorCode = intent.getIntExtra(BluetoothLeScanner.EXTRA_ERROR_CODE, -1); // e.g.  ScanCallback.SCAN_FAILED_INTERNAL_ERROR
                 if (errorCode != -1) {
                     LogManager.w(TAG, "Passive background scan failed.  Code; "+errorCode);